Stillingstittelvelger med predefinerte kategorier Co-authored-by: Mats Bolstad <mats.bolstad@nav.no> Co-authored-by: Håvard Myrvold Johannessen <havard.myrvold.johannessen@nav.no> Co-authored-by: Millad Dagdoni <millad.dagdoni@nav.no>
